+++ src/main/java/org/apiphany/ParameterFunction.java
@@ -54,6 +54,7 @@ public interface ParameterFunction extends Consumer<Map<String, String>> {
 	 * Creates a {@link ParameterFunction} for a single key-value pair, where the value is converted to a string.
 	 *
 	 * @param <T> the type of the value
+	 *
 	 * @param name the parameter name
 	 * @param value the parameter value
 	 * @return a {@link ParameterFunction} that inserts the key-value pair into the map
@@ -68,6 +69,7 @@ public interface ParameterFunction extends Consumer<Map<String, String>> {
 	 *
 	 * @param <T> the type of the key
 	 * @param <U> the type of the value
+	 *
 	 * @param name the parameter name
 	 * @param value the parameter value
 	 * @return a {@link ParameterFunction} that inserts the key-value pair into the map
@@ -80,6 +82,7 @@ public interface ParameterFunction extends Consumer<Map<String, String>> {
 	 * Creates a {@link ParameterFunction} for a single key-value pair, where the value is provided by a supplier.
 	 *
 	 * @param <T> the type of the value
+	 *
 	 * @param name the parameter name
 	 * @param value the supplier of the parameter value
 	 * @return a {@link ParameterFunction} that inserts the key-value pair into the map
@@ -93,6 +96,7 @@ public interface ParameterFunction extends Consumer<Map<String, String>> {
 	 *
 	 * @param <T> the type of the name
 	 * @param <U> the type of the value
+	 *
 	 * @param name the parameter name
 	 * @param value the supplier of the parameter value
 	 * @return a {@link ParameterFunction} that inserts the key-value pair into the map
@@ -107,6 +111,7 @@ public interface ParameterFunction extends Consumer<Map<String, String>> {
 	 *
 	 * @param <T> the type of the key
 	 * @param <U> the type of the value
+	 *
 	 * @param name the supplier of the parameter name
 	 * @param value the supplier of the parameter value
 	 * @return a {@link ParameterFunction} that inserts the key-value pair into the map
@@ -135,6 +140,7 @@ public interface ParameterFunction extends Consumer<Map<String, String>> {
 	 *
 	 * @param <T> the type of the key
 	 * @param <U> the type of the elements
+	 *
 	 * @param name the parameter name
 	 * @param elements the list of elements to join
 	 * @return a {@link ParameterFunction} that inserts the key-value pair into the map
@@ -219,6 +225,7 @@ public interface ParameterFunction extends Consumer<Map<String, String>> {
 	 * Creates a {@link ParameterFunction} that conditionally inserts parameters if the provided object is non-null.
 	 *
 	 * @param <T> the type of the object
+	 *
 	 * @param obj the object to check for null
 	 * @param paramFunctions the {@link ParameterFunction}s to execute if the object is non-null
 	 * @return a {@link ParameterFunction} that conditionally inserts parameters
@@ -231,6 +238,7 @@ public interface ParameterFunction extends Consumer<Map<String, String>> {
 	 * Creates a {@link ParameterFunction} that conditionally inserts parameters based on a predicate.
 	 *
 	 * @param <T> the type of the object
+	 *
 	 * @param obj the object to evaluate
 	 * @param predicate the predicate to test
 	 * @param paramFunctions the {@link ParameterFunction}s to execute if the predicate is true
@@ -245,6 +253,7 @@ public interface ParameterFunction extends Consumer<Map<String, String>> {
 	 * {@link #withPredicate(Object, Predicate, ParameterFunction...)}.
 	 *
 	 * @param <T> the type of the object
+	 *
 	 * @param obj the object to evaluate
 	 * @param predicate the predicate to test
 	 * @param paramFunctions the {@link ParameterFunction}s to execute if the predicate is true